requirements.txt
anyio==4.11.0
certifi==2025.10.5
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
python-telegram-bot==22.5
sniffio==1.3.1
typing_extensions==4.15.0
openai>=1.0.0
reportlab>=4.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
